Added full support to binary read/write for configurations
Now the type "Binary" is allowed for input configurations.
A test configuration is provided conf_04040408.bin in tests directory
This contains the same configuration as the corresponding .txt file
File test_IO.xml is an example of usage.

Support for ILDG data format is almost there but not provided yet